Add exceptions to AppointmentService

+++ BLL/Services/AppointmentService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using BLL.DTO;
+using BLL.Exceptions;
 using BLL.Interfaces;
 using DAL.Entities;
 using DAL.Interfaces;
@@ -23,6 +24,16 @@ namespace BLL.Services
 
         public async Task<AppointmentDTO> CreateAppointment(AppointmentDTO appointmentDTO)
         {
+            var doctor = await _unitOfWork.DoctorRepository.GetByIdAsync(appointmentDTO.DoctorId);
+            var patient = await _unitOfWork.PatientRepository.GetByIdAsync(appointmentDTO.PatientId);
+            if (doctor == null)
+            {
+                throw new EntityNotFoundException(nameof(doctor), appointmentDTO.DoctorId);
+            }
+            if(patient == null)
+            {
+                throw new EntityNotFoundException(nameof(patient), appointmentDTO.PatientId);
+            }
             var appointment = _mapper.Map<Appointment>(appointmentDTO);
             var result = _unitOfWork.AppointmentRepository.Insert(appointment);
             await _unitOfWork.SaveAsync();
@@ -32,30 +43,61 @@ namespace BLL.Services
         public async Task DeleteAppointment(int id)
         {
             var appointment = await _unitOfWork.AppointmentRepository.GetByIdAsync(id);
+            if(appointment == null)
+            {
+                throw new EntityNotFoundException(nameof(appointment), id);
+            }
             _unitOfWork.AppointmentRepository.Delete(appointment);
             await _unitOfWork.SaveAsync();
         }
 
         public async Task<IEnumerable<AppointmentDTO>> GetAllAppointments()
         {
-            return _mapper.Map<IEnumerable<Appointment>, IEnumerable<AppointmentDTO>>(await _unitOfWork.AppointmentRepository.GetAllAsync());
+            var appointments = await _unitOfWork.AppointmentRepository.GetAllAsync();
+            if(appointments == null)
+            {
+                throw new EntityNotFoundException(nameof(appointments), 0);
+            }
+            return _mapper.Map<IEnumerable<Appointment>, IEnumerable<AppointmentDTO>>(appointments);
         }
 
         public async Task<IEnumerable<AppointmentDTO>> GetAllAppointmentsByDoctorId(int id)
         {
+            var doctor = await _unitOfWork.DoctorRepository.GetByIdAsync(id);
+            if(doctor == null)
+            {
+                throw new EntityNotFoundException(nameof(doctor), id);
+            }
             var appointments = await _unitOfWork.AppointmentRepository.GetAllAsync(x => x.DoctorId == id);
+            if(appointments == null)
+            {
+                throw new EntityNotFoundException(nameof(appointments), id);
+            }
             return _mapper.Map<IEnumerable<Appointment>, IEnumerable<AppointmentDTO>>(appointments);
         }
 
         public async Task<IEnumerable<AppointmentDTO>> GetAllAppointmentsByPatientId(int id)
         {
+            var patient = await _unitOfWork.PatientRepository.GetByIdAsync(id);
+            if (patient == null)
+            {
+                throw new EntityNotFoundException(nameof(patient), id);
+            }
             var appointments = await _unitOfWork.AppointmentRepository.GetAllAsync(x => x.PatientId == id);
+            if (appointments == null)
+            {
+                throw new EntityNotFoundException(nameof(appointments), id);
+            }
             return _mapper.Map<IEnumerable<Appointment>, IEnumerable<AppointmentDTO>>(appointments);
         }
 
         public async Task<AppointmentDTO> GetAppointmentById(int id)
         {
             var appointment = await _unitOfWork.AppointmentRepository.GetByIdAsync(id);
+            if(appointment == null)
+            {
+                throw new EntityNotFoundException(nameof(appointment), id);
+            }
             var appointmentDTO = _mapper.Map<AppointmentDTO>(appointment);
             return appointmentDTO;
         }
@@ -63,6 +105,10 @@ namespace BLL.Services
         public async Task<AppointmentDTO> UpdateAppointment(int id, AppointmentDTO appointmentDTO)
         {
             var appointment = await _unitOfWork.AppointmentRepository.GetByIdAsync(id);
+            if (appointment == null)
+            {
+                throw new EntityNotFoundException(nameof(appointment), id);
+            }
 
             appointment.Date = appointmentDTO.Date;
             appointment.DoctorId = appointmentDTO.DoctorId;
